moved installation commands to orchiestrator
